requirements.txt
pydantic
requests
pillow
python-dotenv
streamlit